switch process_revel_scores from bgzip (htslib) to gzip
Use `gzip` instead of `bgzip` in `rule process_revel_scores:`, as `gzip` is commonly available in Linux, whereas `bgzip` is only available in the `htslib`. The alternative would be to add a conda environment to this rule, that installs `htslib`, but as all other commands are plain Linux tools, I would opt for the change proposed here. Also, I double-checked that the flag `-c` means the same thing in both (write to `stdout`).